Introducing the Distinct Sieve
While finding distinct values requires to sorts, first on the images of the functions, then on
the preimages, the generic polymorphic distinct sieve is in keeping with the sort once and
forever hold your peace function design. The distinct sieve sorts on the images of the function
and then flags the topological boundaries in the preimage where the distinct values of the image
start and end. The output is a list of indexes along with the boolean flag indicating whether
the topological boundary is distinct. Secondary functions like sum and avg would then call this
function in place of the call to generate indexes, and then continue with normal processing,
except the push-pop operations will only occur on distinct boundaries.

Also included are minor type casting fixes to the calls to the generic count function