Overhaul Distinct Index Sieve
After extensive debugging a, hopefully correct, implementation of a sieve
to find distinct push/pop boundaries for a give value of the image of
simple measurable functions. Briefly:

Within a partition of a single value in the image sort in the usual topological
ordering, then count the cumulative pushes and pops of the value as topological
boundaries are crossed. Keep the pop boundaries where the cumulative popped equals
the cumulative pushed, and the pushed bounaries where the cumulative pushed equals
the 1 more than the cumulative popped, with the added wrinkle to sieve out
degenerate pushed/popped, that is pushed popped occuring at the exact same topological
point (to be kept either the preimage value must be different or the topology must be
different)