Improved Sorting and Boundary Sifting
Removed fall back sorting of boundaries based on array index, replace with
fall back sorting on the order the functions were aggregated. On the push
side of boundaries sort in the ascending order of the function list, and on
the pop side of the boundary sort in reverse order. This is being introduced
so that the eventual array and string aggregating has the intuitively and
mathematically correct behavior or respect function precedence. This also
introduces an additional field identifying the source function, which will be
used in the array and string aggregation.

Completed the re-factor of te boundary sifting so that sum and average are sifted
after calculation, not in the intial zeroth, first, and second order sum
accumulation.

To-do: test this beast.